+++ application/System/Controllers/PersistentStorageController/PersistentData/HeaterPersistentConfig.hpp
@@ -4,6 +4,6 @@
 
 namespace ATC {
 struct HeaterPersistentConfig {
-    uint16_t targetTemperatureInCelsius;
+    float targetTemperatureInCelsius;
 };
 }
